Day41

+++ src/com/rupali/Day37/arithmetic.java
@@ -2,8 +2,8 @@ package com.rupali.Day37;
 
 public class arithmetic {
     public static void main(String[] var0) {
-        byte var1 = 10;
-        byte var2 = 20;
+        int var1 = 10;
+        int var2 = 20;
         System.out.println(var1 + var2);
         System.out.println(var1 - var2);
         System.out.println(var1 * var2);
